+++ pkg/network/cache/BUILD.bazel
@@ -4,9 +4,10 @@ go_library(
     name = "go_default_library",
     srcs = [
         "cache.go",
+        "dhcpconfig.go",
+        "domaininterface.go",
         "infocache.go",
         "podinterface.go",
-        "types.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/network/cache",
     visibility = ["//visibility:public"],
@@ -25,9 +26,9 @@ go_test(
     srcs = [
         "cache_suite_test.go",
         "cache_test.go",
-        "infocache_test.go",
+        "dhcpconfig_test.go",
+        "domaininterface_test.go",
         "podinterface_test.go",
-        "types_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
