+++ pkg/network/cache/dhcpconfig_test.go
